+++ src/main/java/domain/piece/pawn/Pawn.java
@@ -1,8 +1,9 @@
-package domain.piece;
+package domain.piece.pawn;
 
 import domain.Direction;
 import domain.Square;
 import domain.Team;
+import domain.piece.Piece;
 
 import java.util.List;
 import java.util.Map;
