+++ src/main/java/dto/ChessBoardDTO.java
@@ -3,10 +3,10 @@ package dto;
 import domain.File;
 import domain.Rank;
 import domain.Square;
-import domain.piece.Pawn;
 import domain.piece.Piece;
 import domain.piece.jumping.King;
 import domain.piece.jumping.Knight;
+import domain.piece.pawn.Pawn;
 import domain.piece.sliding.Bishop;
 import domain.piece.sliding.Queen;
 import domain.piece.sliding.Rook;
